+++ share/p2p/shrexpush/pubsub_test.go
@@ -1,4 +1,4 @@
-package push
+package shrexpush
 
 import (
 	"context"
