+++ share/p2p/shrexpush/subscription.go
@@ -1,4 +1,4 @@
-package push
+package shrexpush
 
 import (
 	"context"
